Fixing CI tests.
Setting CIRCLE_TEST_REPORTS environment variable even on CircleCI 2.0 because tozd/meteor-test-runner uses it to know that it has to export xunit results. We are not using it, but maybe we will in the future.

We do not need to run xvfb-run because chrome runs in headless mode by default.